refactor: ElementWise + Embedding own analytical mem-op latency (AIC-477)
Phase 2.1 of the SDK refactor. ElementWise and Embedding move out of
``_legacy.py`` into per-class modules under ``operations/``. Neither has
CSV-backed data, so the base ``Operation.load_data`` no-op default
applies — no ``_data_cache``, no SOL correction, no extrapolation.

Their analytical mem-op latency formula is extracted from
``PerfDatabase.query_mem_op`` into a pure
``interpolation.estimate_mem_op(gpu_spec, mem_bytes, database_mode)``
function. ``query_mem_op`` becomes a thin LRU-cached delegating wrapper.
ElementWise and Embedding ``query()`` paths now call ``estimate_mem_op``
directly, eliminating one of the cross-cutting ``database.*`` couplings.

ISSUE-16 retires ``PerfDatabase.query_mem_op`` once the remaining callers
(ContextAttention QK-norm/RoPE/KV-write paths in ISSUE-06, Mamba2 and
Mamba2Kernel SOL fallback in ISSUE-09) have migrated.

Behavior preservation:
- ``query_mem_op`` callers (3 unit tests in test_moe_mla.py +
  test_database_helpers.py) keep exercising the LRU-cached wrapper —
  unchanged interface, delegates to the new pure function.
- Lazy field lookup preserved: SOL paths only touch ``mem_bw``;
  EMPIRICAL / SILICON / HYBRID also read
  ``mem_bw_empirical_scaling_factor`` + ``mem_empirical_constant_latency``.

Verification:
- 670 unit tests pass + 1 skipped (no regression).
- 38 regression snapshots pass + 6 skipped (same as post-rebase baseline).
- ruff check + format clean.
