+++ src/aiconfigurator/sdk/operations/_legacy.py
@@ -1060,84 +1060,6 @@ class MLABmm(Operation):
         return self._weights * self._scale_factor
 
 
-class Embedding(Operation):
-    """
-    Embedding operation.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        row_size: int,
-        column_size: int,
-        empirical_bw_scaling_factor: float = 0.3,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._row_size = row_size
-        self._column_size = column_size
-        self._weights = row_size * column_size * 2
-        self._empirical_bw_scaling_factor = empirical_bw_scaling_factor
-        self._constant_latency = 5e-6  # 5us
-
-    # sol only
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """Query embedding latency with power data."""
-        x = kwargs.get("x")
-        d2d_bytes = x * self._column_size * 2
-
-        result = database.query_mem_op(d2d_bytes)
-        return PerformanceResult(
-            float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
-class ElementWise(Operation):
-    """
-    Element-wise operation.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        dim_in: int,
-        dim_out: int,
-        empirical_bw_scaling_factor: float = 0.8,
-        **kwargs,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._weights = 0.0
-        self._empirical_bw_scaling_factor = empirical_bw_scaling_factor
-        self._constant_latency = 5e-6  # 5us
-        self._dim_in = dim_in
-        self._dim_out = dim_out
-        self._scale_num_tokens = kwargs.get("scale_num_tokens", 1)
-
-    # sol only
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """Query element-wise operation latency with power data."""
-        x = kwargs.get("x")  # num tokens
-        x //= self._scale_num_tokens
-        read_bytes = x * self._dim_in * 2  # bfloat16 for act
-        write_bytes = x * self._dim_out * 2
-
-        result = database.query_mem_op(read_bytes + write_bytes)
-        return PerformanceResult(
-            float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
 class WideEPGenerationMLA(Operation):
     """
     WideEP Generation MLA operation.
