+++ src/aiconfigurator/sdk/perf_database.py
@@ -5751,31 +5751,18 @@ class PerfDatabase:
     ) -> PerformanceResult | tuple[float, float, float]:
         """Query memory-operation latency analytically (no CSV data).
 
+        Thin LRU-cached wrapper around ``interpolation.estimate_mem_op``.
+        ISSUE-16 retires this wrapper once every caller (Embedding,
+        ElementWise, ContextAttention QK-norm/RoPE/KV-write paths, Mamba2,
+        Mamba2Kernel SOL fallback) has migrated to the pure function.
+
         Returns:
             PerformanceResult acting as float (latency in ms); energy via ``.energy``.
             For SOL_FULL, returns a ``(sol_time, 0, sol_time)`` tuple.
         """
-        gpu_spec = self.system_spec["gpu"]
-
-        def get_sol() -> tuple[float, float, float]:
-            sol_time = mem_bytes / gpu_spec["mem_bw"] * 1000
-            return sol_time, 0, sol_time
-
-        def get_empirical() -> float:
-            return (
-                mem_bytes / (gpu_spec["mem_bw"] * gpu_spec["mem_bw_empirical_scaling_factor"])
-                + gpu_spec["mem_empirical_constant_latency"]
-            ) * 1000
-
         if database_mode is None:
             database_mode = self._default_database_mode
-        if database_mode == common.DatabaseMode.SOL:
-            return PerformanceResult(get_sol()[0], energy=0.0, source="sol")
-        if database_mode == common.DatabaseMode.SOL_FULL:
-            return get_sol()
-        # EMPIRICAL / SILICON / HYBRID share the same empirical formula. There is
-        # no silicon table for raw memory ops, so always tag as ``empirical``.
-        return PerformanceResult(get_empirical(), energy=0.0, source="empirical")
+        return interpolation.estimate_mem_op(self.system_spec["gpu"], mem_bytes, database_mode)
 
     def query_mamba2(
         self,
